added cohort system and certificating system. added recaptcha

+++ routes/api.php
@@ -16,6 +16,8 @@ use App\Http\Controllers\AuthController;
 use App\Http\Controllers\PasswordController;
 use App\Http\Controllers\CartController;
 use App\Http\Controllers\SaleController;
+use App\Http\Controllers\CohortController;
+use App\Http\Controllers\CertificateController;
 use App\Http\Controllers\AffiliateController;
 use App\Http\Controllers\FulfillmentController;
 use Illuminate\Support\Facades\DB;
@@ -58,7 +60,7 @@ Route::get('/certification-course/{course_code}', [CertificationCourseController
 Route::middleware(['should_use:admin-jwt', 'jwt_auth'])->group(function () {
     Route::post('/admin/offshore-course/create', [OffshoreCourseController::class, 'store']);
     Route::post('/admin/offshore-course/{course_title}/edit', [OffshoreCourseController::class, 'edit']);
-    Route::delete('/offshore-course/{course_title}/delete', [OffshoreCourseController::class, 'delete']);
+    Route::delete('/admin/offshore-course/{course_title}/delete', [OffshoreCourseController::class, 'delete']);
 });
 Route::get('/offshore-course/names', [OffshoreCourseController::class, 'get_names']);
 Route::get('/offshore-courses/{count}', [OffshoreCourseController::class, 'get_list']);
@@ -70,6 +72,7 @@ Route::middleware(['should_use:admin-jwt', 'jwt_auth'])->group(function () {
     Route::get('/admin/courses', [CourseController::class, 'get']); 
     Route::get('/admin/resources', [ResourceController::class, 'get']);
 });
+Route::get('/{type}/enrolled-students/{course_identity}', [CourseController::class, 'get_enrolled_students']);
 
 
 
@@ -117,6 +120,31 @@ Route::middleware(['should_use:admin-jwt', 'jwt_auth'])->group(function () {
     Route::get('/admin/sales', [SaleController::class, 'get_all']);
 });
 
+
+// routes for cohorts
+Route::middleware(['should_use:admin-jwt', 'jwt_auth'])->group(function () {
+    Route::post('/admin/cohort', [CohortController::class, 'store']);
+    Route::get('/admin/cohort/{name}/start', [CohortController::class, 'start']);
+    Route::get('/admin/cohort/{name}/conclude', [CohortController::class, 'conclude']);
+    Route::get('/admin/cohort/{name}/abort', [CohortController::class, 'abort']);
+    Route::get('/admin/cohort/{name}/delete', [CohortController::class, 'delete']);
+    Route::get('/admin/cohorts', [CohortController::class, 'get_all']);
+    Route::get('/admin/cohort/names', [CohortController::class, 'get_names']);
+    Route::get('/admin/cohort/{name}', [CohortController::class, 'get']);
+    Route::post('/admin/cohort/{name}/notify-students', [CohortController::class, 'notify_students']);
+    Route::get('/admin/cohort/{name}/students', [CohortController::class, 'all_students_showing_those_in_cohort']);
+    Route::post('/admin/cohort/{name}/students', [CohortController::class, 'add_students']);
+    Route::get('/admin/cohort/{name}/certificates', [CohortController::class, 'get_students_certificates']);
+    Route::get('/admin/cohort/{name}/edit', [CohortController::class, 'get_cohort_for_edit']);
+    Route::post('/admin/cohort/{name}/edit', [CohortController::class, 'edit']);
+});
+
+// routes for certificates
+Route::middleware(['should_use:admin-jwt', 'jwt_auth'])->group(function () {
+    Route::post('/admin/certificates', [CertificateController::class, 'upload']);
+});
+Route::get('/certificate/{type}', [CertificateController::class, 'get']);
+
 // routes for fulfillments
 Route::middleware(['should_use:admin-jwt', 'jwt_auth'])->group(function () {
     Route::get('/admin/fulfillment/{id}', [FulfillmentController::class, 'get']);
@@ -134,6 +162,15 @@ Route::middleware(['should_use:admin-jwt', 'jwt_auth'])->group(function () {
     Route::get('/admin/requests/{count}', [ContactRequestController::class, 'get_list']);
 });
 
+// routes for users and affiliates
+
+Route::middleware(['should_use:admin-jwt', 'jwt_auth'])->group(function () {
+    Route::get('/admin/users', [StudentController::class, 'get_all']);
+    Route::get('/admin/affiliates', [AffiliateController::class, 'get_all']);
+    Route::get('/admin/user/{gmail}', [StudentController::class, 'get']);
+    Route::get('/admin/affiliate/{gmail}', [AffiliateController::class, 'get']);
+});
+
 
 
 // routes for students
@@ -144,20 +181,28 @@ Route::post('/student/send-otp', [StudentController::class, 'send_otp']); //send
 Route::middleware(['should_use:student-jwt', 'jwt_auth'])->group(function () {
     Route::post('/student/cart', [CartController::class, 'add']);
     Route::get('/student/cart', [StudentController::class, 'get_cart']);
-    Route::post('/student/get_carted_course', [StudentController::class, 'get_carted_course']);
+    Route::post('/student/remove-from-cart', [CartController::class, 'remove']);
+    Route::get('/student/courses', [StudentController::class, 'get_enrolled_courses']);
+    Route::post('/student/get_carted_course', [StudentController::class, 'get_course']);
+    Route::post('/student/get_enrolled_course', [StudentController::class, 'get_course']);
     Route::post('/student/watchlist', [StudentController::class, 'add_event_to_watchlist']);
     Route::get('/student/watchlist', [StudentController::class, 'get_event_watchlist']);
     Route::get('/student/watchlist/{event}', [StudentController::class, 'get_event_watchlist_event']);
     Route::get('/student/profile', [StudentController::class, 'get_profile']);
     Route::post('/student/profile', [StudentController::class, 'update_profile']);
     Route::get('/student/become-affiliate', [AffiliateController::class, 'create_referral_code']);
-    Route::get('/student/renew-referral-code', [AffiliateController::class, 'create_referral_code']);
+    Route::get('/student/renew-referral-code', [AffiliateController::class, 'renew_referral_code']);
     Route::get('/student/load-affiliate-portal', [AffiliateController::class, 'load_affiliate_portal']);
     Route::post('/student/payout', [FulfillmentController::class, 'add']);
     Route::get('/student/trending-courses', [CourseController::class, 'trending_courses']);
+    Route::get('/student/certificates', [CertificateController::class, 'get_my_certificates']);
+    Route::get('/student/download-certificate', [CertificateController::class, 'download_certificate']);
 });
 Route::get('/student/{phone_number}', [StudentController::class, 'get']);
 Route::get('/students/{count}', [StudentController::class, 'get_list']);
+
+
+
 
 
 
@@ -170,6 +215,16 @@ Route::post('/login/resend-otp/{type}/{email}', [AuthController::class, 'resend_
 Route::get('/test-auth', function() {
     return response()->json(['user'=> auth()->guard('student-jwt')->user(), 'message'=>bcrypt('mlaadmin12345')], 200);
 })->middleware(['should_use:student-jwt', 'jwt_auth']);
+
+
+
+// routes for password confirmation
+Route::post('/admin/confirm-password', [AuthController::class, 'confirm_admin_password'])
+->middleware(['should_use:admin-jwt', 'jwt_auth']);; //confirms authenticated admin's password
+
+Route::post('/student/confirm-password', [AuthController::class, 'confirm_student_password'])
+->middleware(['should_use:student-jwt', 'jwt_auth']);; //confirms authenticated student's password
+
 
 
 //routes for password reset
@@ -213,5 +268,5 @@ Route::get('/admin-dashboard', function (Request $request) {
 Route::get('/admin/fetch_student_name/{email}', [StudentController::class, 'fetch_name'])->middleware(['should_use:admin-jwt', 'jwt_auth']);
 
 // routes for affiliate
-Route::get('/admin/fetch-affiliate/{code}', [AffiliateController::class, 'fetch_affiliate'])->middleware(['should_use:admin-jwt', 'jwt_auth']);
+Route::get('/fetch-affiliate/{code}', [AffiliateController::class, 'fetch_affiliate']);#->middleware(['should_use:admin-jwt', 'jwt_auth']);
